refactor: extract platform/asset helpers from core.py into _platform.py
